Added learning threshold files and mini_imagenet episodic trained

+++ TestFewShotNoveltyModel.py
@@ -11,16 +11,11 @@ import numpy as np
 import pandas as pd
 import torch
 import argparse
-#from torch import nn
 from torch.utils.data import DataLoader
 
-#from easyfsl.datasets import CUB
-#from easyfsl.datasets import MiniImageNet
-#from easyfsl.datasets import EasySet
-#from easyfsl.modules import resnet12
 from PrototypicalNetworksNovelty import PrototypicalNetworksNovelty
 from utilsNovelty import evaluate
-from NoveltyThreshold import getLearnedThreshold, BayesTwoClassThreshold, StdTimesTwoThredshold
+from NoveltyThreshold import StdTimesTwoThredshold
 
 from easyfsl.modules import resnet12
 from easyfsl.methods import PrototypicalNetworks, RelationNetworks, MatchingNetworks, TransductiveFinetuning
@@ -33,30 +28,42 @@ from torchvision.models import resnet50 #, ResNet50_Weights
 from torchvision.models import resnet34 #, ResNet34_Weights
 from torchvision.models import resnet18 #, ResNet18_Weights
 
-modelPth = "_model" # First model trained classic and pretrained weights
-#modelPth = "_classic_pretrained" # Improved models with pretrained weights and classic training
+
+def get_train_method(argsWeights):
+    
+    trainMethod = "_model" # First model trained classic and pretrained weights
+    #trainMethod = "_classic_pretrained" # Improved models with pretrained weights and classic training
+    if argsWeights == 'mini_imagenet':
+        trainMethod = '_episodic' # Mini imagenet trained from scratch episodic
+    if argsWeights == 'CUB': 
+        trainMethod = '_classic_pretrained' # Best performing models on CUB dataset  
+    
+    return trainMethod
+    
 
 def load_model(argsModel, argsWeights):
 
+    trainMethod = get_train_method(argsWeights)
+  
     if argsModel == 'resnet50':
         print('resnet50')
         #ResNetModel = resnet50(weights=ResNet50_Weights.IMAGENET1K_V2) # 80.86, 25.6M
         ResNetModel = resnet50(pretrained=True) # 80.86, 25.6M
-        model = EmbeddingsModel(ResNetModel, num_classes, use_fc=False)
-        modelName = "./models/Resnet50_"+argsWeights+modelPth+".pth"
+        model = EmbeddingsModel(ResNetModel, num_classes, use_fc=False)        
+        modelName = "./models/Resnet50_"+argsWeights+trainMethod+".pth"
         feat_dim = 2048
     if argsModel == 'resnet34':
         print('resnet34')
         ResNetModel = resnet34(pretrained=True) # 80.86, 25.6M
         model = EmbeddingsModel(ResNetModel, num_classes, use_fc=False)
-        modelName = "./models/Resnet34_"+argsWeights+modelPth+".pth"
+        modelName = "./models/Resnet34_"+argsWeights+trainMethod+".pth"
         feat_dim = 512
     if argsModel == 'resnet18':
         print('resnet18')
         #ResNetModel = resnet18(weights=ResNet18_Weights.IMAGENET1K_V1) 
         ResNetModel = resnet18(pretrained=True) # 80.86, 25.6M
         model = EmbeddingsModel(ResNetModel, num_classes, use_fc=False)
-        modelName = "./models/Resnet18_"+argsWeights+modelPth+".pth"
+        modelName = "./models/Resnet18_"+argsWeights+trainMethod+".pth"
         feat_dim = 512
     if argsModel == 'resnet12':
         print('resnet12')
@@ -72,11 +79,6 @@ def load_model(argsModel, argsWeights):
         #ResNetModel.load_state_dict(modelSaved.state_dict())
         model.load_state_dict(modelSaved.state_dict())
  
-    subDir = args.weights + '/'
-    if os.path.exists(resDir+subDir) == False:
-        os.mkdir(resDir+subDir)
-        print("Create result directory", resDir+subDir)
-
     model.eval()
     model = model.to(DEVICE)
     
@@ -116,15 +118,15 @@ def load_test_dataset(argsDataset, argsLearning):
             print("miniImageNet Val dataset")
         else:
             test_set = FewShotDataset(split="test", image_size=image_size, root=dataDirMiniImageNet, training=False)
-            print("miniImageNet Val dataset")
+            print("miniImageNet Test dataset")
             
     return test_set
 
 
-def get_threshold_learned(argsDataset, argsModel, nameNoveltyLearned, n_shot, useBayesThreshold=True):
+def get_threshold_learned(argsModel, argsWeights, nameNoveltyLearned, n_shot, useBayesThreshold=True):
     
-    noveltyLearnedFile = "./learned/" + argsDataset + '_' + argsModel + nameNoveltyLearned + '.csv'
-    print(noveltyLearnedFile)
+    noveltyLearnedFile = "./learned/" + argsModel + '_' + argsWeights + nameNoveltyLearned + '.csv'
+    print("Learned threshold file", noveltyLearnedFile)
     
     df = pd.read_csv(noveltyLearnedFile)
     
@@ -136,6 +138,7 @@ def get_threshold_learned(argsDataset, argsModel, nameNoveltyLearned, n_shot, us
         threshold = StdTimesTwoThredshold(df['Average'].to_numpy()[0], df['Std'].to_numpy()[0])
         
     return threshold
+
 
 def test_or_learn(test_set, test_sampler, few_shot_classifier, 
                   novelty_th, use_novelty, learn_th, n_workers, DEVICE):
@@ -161,18 +164,18 @@ def test_or_learn(test_set, test_sampler, few_shot_classifier,
 if __name__=='__main__':
 
     parser = argparse.ArgumentParser()
-    parser.add_argument('--model', default='resnet18') #resnet12 (Omniglot), resnet18, resnet34, resnet50
-    parser.add_argument('--weights', default='euMoths') #ImageNet, euMoths, CUB, Omniglot
-    parser.add_argument('--dataset', default='euMoths') #miniImagenet, euMoths, CUB, Omniglot
+    parser.add_argument('--model', default='resnet50') #resnet12 (Omniglot), resnet18, resnet34, resnet50
+    parser.add_argument('--weights', default='mini_imagenet') #ImageNet, mini_imagenet, euMoths, CUB, Omniglot
+    parser.add_argument('--dataset', default='miniImagenet') #miniImagenet, euMoths, CUB, Omniglot
     # parser.add_argument('--model', default='resnet12') #resnet12 (Omniglot), resnet18, resnet34, resnet50
     # parser.add_argument('--weights', default='Omniglot') #ImageNet, euMoths, CUB, Omniglot
     # parser.add_argument('--dataset', default='Omniglot') #miniImagenet, euMoths, CUB, Omniglot
     parser.add_argument('--novelty', default='True', type=bool) #default false when no parameter - automatic False when learning True
-    parser.add_argument('--learning', default='True', type=bool) #default false when no parameter - learn threshold for novelty detection
+    parser.add_argument('--learning', default='', type=bool) #default false when no parameter - learn threshold for novelty detection
     parser.add_argument('--shot', default=5, type=int) 
     parser.add_argument('--way', default=6, type=int) # Way 0 is novelty class
     parser.add_argument('--query', default=6, type=int)
-    parser.add_argument('--threshold', default='std') # bayes or std threshold to be used
+    parser.add_argument('--threshold', default='bayes') # bayes or std threshold to be used
     args = parser.parse_args()
   
     resDir = "./result/"
@@ -216,13 +219,14 @@ if __name__=='__main__':
         num_classes = 140  
     if args.weights == 'Omniglot':
         num_classes = 3856  
+    if args.weights == 'mini_imagenet':
+        num_classes = 60
         
     if args.learning:
         n_test_tasks = 50 # 50 learning on validation
     else:
-        n_test_tasks = 100 # 500 test
+        n_test_tasks = 10 # 500 test
         
-   
     #%% Create model and prepare for training
     #DEVICE = "cuda"
     DEVICE = "cpu"
@@ -258,11 +262,16 @@ if __name__=='__main__':
     
     test_set = load_test_dataset(args.dataset, args.learning)
     
-    
+    #subDir = args.weights + '/'
+    subDir =  'test/'
+    if os.path.exists(resDir+subDir) == False:
+        os.mkdir(resDir+subDir)
+        print("Create result directory", resDir+subDir)
+   
     #test(model, test_set, test_sampler, few_shot_classifier, n_workers)
     if args.learning:     
-        resFileName = args.dataset + '_' + args.model + "_novelty_learned.csv"
-        resFile = open(resDir+subDir+resFileName, "a")
+        resFileName = args.model + '_' + args.weights + "_novelty_learned.csv"
+        resFile = open(resDir+subDir+resFileName, "w")
         
         few_shot = few_shot_classifiers[0]
         print(few_shot[0])
@@ -273,8 +282,11 @@ if __name__=='__main__':
         print(line)
         resFile.write(line)   
         
-        #for n_shot in [1,2,3,4,5,6,7,8,9]: # Learn distribution for each shot in range 1-9
-        for n_shot in [1,2,3,4,5]: # Learn distribution for each shot in range 1-9
+        shotRange = [1,2,3,4,5,6,7,8,9] # Learn distribution for each shot in range 1-9
+        if args.dataset == "euMoths":
+            shotRange = [1,2,3,4,5] # Learn distribution for each shot in range 1-5, euMoths only
+
+        for n_shot in shotRange:
         
             test_sampler = TaskSampler(
                 test_set, n_way=n_way, n_shot=n_shot, n_query=n_query, n_tasks=n_test_tasks
@@ -292,18 +304,27 @@ if __name__=='__main__':
         print("Result saved to", resFileName)
             
     else: #Testing
-        resFileName = args.dataset + '_' + args.model + '_' + args.threshold + '_' + str(n_way) + 'way_' + str(n_shot) +"shot_novelty_test.txt"
-        resFile = open(resDir+subDir+resFileName, "w")
 
         test_sampler = TaskSampler(
             test_set, n_way=n_way, n_shot=n_shot, n_query=n_query, n_tasks=n_test_tasks
         )
         
-        novelty_th = getLearnedThreshold(args.weights, args.model, args.shot) # Old method 
-        #novelty_th = get_threshold_learned(args.weights, args.model, modelPth+'_novelty_learned', n_shot, useBayesThreshold=(args.threshold == 'bayes'))
+        #novelty_th = getLearnedThreshold(args.weights, args.model, args.shot) # Old method 
+        novelty_th = get_threshold_learned(args.model, args.weights, get_train_method(args.weights)+'_novelty_learned', n_shot, useBayesThreshold=(args.threshold == 'bayes'))
     
-        line = "FewShotClassifier,Way,Shot,Query,Accuracy,Threshold,Average,Std\n"
-        resFile.write(line)   
+        #resFileName = args.model + '_' + args.dataset + '_' + args.threshold + '_' + str(n_way) + 'way_' + str(n_shot) +"shot_novelty_test.txt"
+        if args.novelty:
+            resFileName = args.model + '_' + args.dataset + '_' + args.threshold + "_novelty.txt" # With novelty detection
+        else:
+            resFileName = args.model + '_' + args.dataset + "_few_shot.txt" # Normal few-shot learning
+            
+        if os.path.exists(resDir+subDir+resFileName):
+            resFile = open(resDir+subDir+resFileName, "a") # Append to existing result file
+        else:
+            resFile = open(resDir+subDir+resFileName, "w") # Create new result file with header           
+            line = "FewShotClassifier,Way,Shot,Query,Accuracy,Threshold\n"
+            resFile.write(line)   
+            
         for few_shot in few_shot_classifiers:
             print(few_shot[0])
             print("Use softmax", few_shot[1].use_softmax)
@@ -311,7 +332,7 @@ if __name__=='__main__':
             accuracy, threshold, avg, std, avg_o, std_o = test_or_learn(test_set, test_sampler, few_shot_classifier, 
                                                                         novelty_th, args.novelty, args.learning, 
                                                                         n_workers, DEVICE)
-            line = few_shot[0] + ',' + str(n_way) + ','  + str(n_shot) + ','  + str(n_query) + ',' + str(accuracy) + ',' + str(threshold) + ',' + str(avg) + ',' + str(std) + '\n'
+            line = few_shot[0] + ',' + str(n_way) + ','  + str(n_shot) + ','  + str(n_query) + ',' + str(accuracy) + ',' + str(threshold) + '\n'
             resFile.write(line)    
         resFile.close()
         print("Result saved to", resFileName)
